+++ lib/commands/list.js
@@ -1,0 +1,6 @@
+const filter = require('lodash').filter;
+
+module.exports = function(currentState, channel) {
+  return filter(currentState, ['channel', channel]);
+};
+
