+++ lib/pairrit.js
@@ -1,5 +1,6 @@
-const _ = require('lodash');
-const SHA256  = require('crypto-js/sha256');
+const extend = require('lodash').extend;
+const filter = require('lodash').filter;
+const SHA256 = require('crypto-js/sha256');
 
 function generateKey(pairChannel, pairName) {
   return SHA256(`${pairChannel}-${pairName}`);
@@ -22,7 +23,7 @@ function updatePair(currentState, options) {
 
   if (newPair !== undefined) {
     newPairState[key] = newPair;
-    return _.extend(newPairState, currentState);
+    return extend(newPairState, currentState);
   } else {
     delete currentState[key];
     return currentState;
@@ -30,39 +31,12 @@ function updatePair(currentState, options) {
 };
 
 function listPairs(currentState, channel) {
-  return _.filter(currentState, ['channel', channel]);
+  return filter(currentState, ['channel', channel]);
 };
-
-function joinOrCreatePair(options) {
-  const newPair = Object.assign(
-    {
-      channel: options.pairChannel,
-      name: options.pairName,
-      participants: new Set()
-    },
-    options.pairState
-  );
-
-  newPair.participants.add(options.userName)
-
-  return newPair;
-};
-
-function leavePair(options) {
-  const newPair = Object.assign({}, options.pairState);
-
-  newPair.participants.delete(options.userName);
-
-  if (newPair.participants.size === 0) {
-    return undefined;
-  } else {
-    return newPair;
-  };
-}
 
 module.exports = {
-  join:    joinOrCreatePair,
-  leave:   leavePair,
-  list:    listPairs,
+  join:    require('./commands/join'),
+  leave:   require('./commands/leave'),
+  list:    require('./commands/list'),
   update:  updatePair
 };
