+++ spec/lib/pairrit_spec.js
@@ -100,105 +100,4 @@ describe('Pairrit', function() {
       expect(result).to.deep.equal(expected);
     });
   });
-
-  describe('listing pairs', function() {
-    function generateKey(pairChannel, pairName) {
-      return SHA256(`${pairChannel}-${pairName}`);
-    };
-
-    it('list all pairs for a given channel', function() {
-      const pairA = {
-        channel: '#gotham',
-        name: 'batcave',
-        participants: new Set(['alfred', 'batman', 'batwoman', 'robin'])
-      }
-
-      const pairB = {
-        channel: '#gotham',
-        name: 'wayne-manor',
-        participants: new Set(['bruce-wayne', 'selina-kyle', 'barbara-gordon', 'kate-kane'])
-      }
-
-      const pairC = {
-        channel: '#xaviers-school-for-gifted-youngsters',
-        name: 'x-men',
-        participants: new Set(['wolverine', 'cyclops'])
-      }
-
-      const currentState = {
-        [generateKey('#gotham', pairA.name)]: pairA,
-        [generateKey('#gotham', pairB.name)]: pairB,
-        [generateKey('#xaviers-school-for-gifted-youngsters', pairC.name)]: pairC
-      }
-
-      const expected = [pairA, pairB];
-
-      const result = pairrit.list(currentState, '#gotham');
-
-      expect(result).to.deep.equal(expected)
-    })
-  });
-  describe('/pair join' , function() {
-    it('creates a new pair with the given user and name',  function() {
-      const options = {
-        userName: 'batman',
-        pairChannel: '#gotham',
-        pairName: 'batcave',
-        pairState: undefined
-      }
-
-      const expected = { channel: '#gotham', name: 'batcave', participants: new Set(['batman']) }
-
-      const result = pairrit.join(options)
-
-      expect(result).to.deep.equal(expected);
-    });
-
-    it('adds a user to an existing pair',  function() {
-      const options = {
-        userName: 'alfred',
-        pairChannel: '#gotham',
-        pairName: 'batcave',
-        pairState: {name: 'batcave', participants: new Set(['batman']) }
-      }
-
-      const expected = {
-        channel: '#gotham',
-        name: 'batcave', 
-        participants: new Set(['batman', 'alfred'])
-      }
-
-      const result = pairrit.join(options)
-
-      expect(result).to.deep.equal(expected);
-    });
-
-    it('updates a pair when someone leaves', function() {
-      const options = {
-        userName: 'batman',
-        pairName: 'batcave',
-        pairState: {name: 'batcave', participants: new Set(['batman', 'alfred'])}
-      }
-
-      const expected = { name: 'batcave', participants: new Set(['alfred']) }
-
-      const result = pairrit.leave(options)
-
-      expect(result).to.deep.equal(expected);
-    });
-
-    it('returns undefined when the last participant leaves', function() {
-      const options = {
-        userName: 'batman',
-        pairName: 'batcave',
-        pairState: {name: 'batcave', participants: new Set(['batman'])}
-      }
-
-      const expected = undefined;
-
-      const result = pairrit.leave(options)
-
-      expect(result).to.equal(expected);
-    });
-  });
 });
